requirements.txt
diffusers
numpy
opencv-python
Pillow
torch
transformers